refactor : 사다리게임 step2 Line객체 생성 책임 관련 리팩토링 객체생성 책임을 객체 스스로가 관리하기 위해 객체생성 메소드를 서비스인 LineGenerator에서 Line 클래스로 이관함. LineGenerator는 삭제 처리함.
