+++ src/main/java/ladderGame/step2/model/Line.java
@@ -1,20 +1,47 @@
 package ladderGame.step2.model;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import ladderGame.step2.service.LineGenerator;
+import java.util.Random;
+import java.util.stream.IntStream;
 
 public class Line {
+
+  private static final int INT_ONE = 1;
+
+  private static Random random = new Random();
 
   private List<Boolean> points;
 
   public Line(final int countOfPerson) {
-    this.points = Collections.unmodifiableList(LineGenerator.createLine(countOfPerson));
+    this.points = Collections.unmodifiableList(createLine(countOfPerson));
   }
 
   public List<Boolean> getPoints() {
     return points;
+  }
+
+  public static List<Boolean> createLine(final int countOfPerson) {
+
+    List<Boolean> lines = new ArrayList<>();
+
+    lines.add(false);
+
+    IntStream.range(INT_ONE, countOfPerson)
+        .forEach(person -> lines.add(isHaveNode(lines.get(person - INT_ONE))));
+
+    return lines;
+  }
+
+  private static Boolean isHaveNode(final Boolean afterNode) {
+
+    if (afterNode) {
+      return false;
+    }
+    return random.nextBoolean();
+
   }
 
   @Override
